Disable refreshLogger and make refreshRootLogger only work on Proxy
