+++ kernel/logging/core/src/main/java/org/apache/shardingsphere/logging/rule/builder/LoggingRuleBuilder.java
@@ -35,7 +35,8 @@ public final class LoggingRuleBuilder implements GlobalRuleBuilder<LoggingRuleCo
     @Override
     public LoggingRule build(final LoggingRuleConfiguration ruleConfig, final Map<String, ShardingSphereDatabase> databases, final ConfigurationProperties props) {
         LoggingUtils.syncLoggingConfig(ruleConfig, props);
-        LoggingUtils.refreshLogger(ruleConfig);
+        // TODO Make it only work on Proxy
+        // LoggingUtils.refreshLogger(ruleConfig);
         return new LoggingRule(ruleConfig);
     }
