+++ mode/type/cluster/core/pom.xml
@@ -48,11 +48,6 @@
             <artifactId>shardingsphere-single-core</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>ch.qos.logback</groupId>
-            <artifactId>logback-classic</artifactId>
-            <scope>compile</scope>
-        </dependency>
         
         <dependency>
             <groupId>org.apache.shardingsphere</groupId>
